+++ NEWS.md
@@ -9,6 +9,17 @@ command-line scripts. It comes with options for working with figure, grid data,
 and horizons. Options can be stored in text files and read passing the path to
 the `-c` flag.
 
+#### Faster `HierarchicalGridData`
+
+The function that finds the component corresponding to a given point in
+`HierarchicalGridData` was significantly sped up by adopting the algorithm used
+in `PostCactus` (developed by Wolfgang Kastaun). This new algorithm is
+significantly faster, but it will not work in case the refinement factors across
+different levels are not constant integers. In that case (e.g., the refinement
+boundaries are at 1, 2, and 6), the older algorithm will be used. For large
+towers of refinement levels and hundreds of MPI processes, the new algorithm is
+orders of magnitude faster.
+
 #### General
 - Added `bins` examples: `save_resampled_grid_data`
 
@@ -19,8 +30,14 @@ the `-c` flag.
 - The function `load_UniformGridData` can now read `.npz` files.
 
 #### Bug fixes
-- Fix bug that, under certain circumstances, resulted in `cactus_grid_function`
+- Fixed bug that, under certain circumstances, resulted in `cactus_grid_function`
   not correctly indexing all the 3D data files
+- Fixed a test that was triggering the wrong error.
+
+#### Breaking changes
+- `finest_level_component_at_point` is now `finest_component_at_point` and
+   returns directly the component as `UniformGridData` as opposed to the
+   refinement level and component number.
 
 ## Version 1.0.0 (11 April 2021)